2.2.0 (#7)
- `Feat`: `HTTPS` redirect
- `Fix`: *Unintended* `visit limits`
- `Change`: Increased `visit limits`
- `Change`: Much better `Disclaimer`. Now provides `Privacy and Cookies policy`
- `Change`: `Login redirect` behaves better
- `Change`: `Player` borders for guide
- `Change`: `Press OK to play` player control instructions
- `Change`: Add `report` as anon/logged user on report
- `Change`: `Top navigation` bar is now powered by `Bootstrap`
- `Change`: `Website Status` sidebar => `Announcement` alert box
- `Change`: Removed `search textbox` on navigation sidebar
- `Change`: favicon changed to `Amaterasu` from `Eros`
- `Change`: `API HTTP statuses` were tweaked
- `Feat`: User auth middleware
- `Feat`:dev: TSLint rule (`template-strings`)
- `Change`: `Views` rewritten
- `Change`:dev: Readme: Added `Pug Task`
- `Feat`: Global `Search textbox`
- `Change`: `Info Table` background `darker colour`
- `Change`: Increased ratelimit for `GET /list`
- `Change`: sliced `GET /search` result length to `ten items`

+++ tools/finalize.js
@@ -7,9 +7,13 @@ const ex = require('util').promisify(exec);
 // tslint:disable-next-line:no-console
 const log = (message, obj) => console.log('finalize: ' + message, obj || '');
 
+const pugTask = process.argv.includes('--pug');
+
 // -- Uglify TypeScript build
 (async () => {
-  const command = 'uglifyjs-folder build/proto --config-file uglify.json -ey -x .js -o build';
+  if (pugTask) return true;
+
+  const command = 'terser-folder build/proto --config-file .terser-folder.json -ey -x .js -o build';
   const { stderr } = await ex(command, { cwd: resolve(__dirname, '..') });
 
   if (stderr) throw err;
@@ -21,6 +25,8 @@ const log = (message, obj) => console.log('finalize: ' + message, obj || '');
 
 // -- Remove TypeScript build
 .then(async () => {
+  if (pugTask) return true;
+
   const proto = resolve(__dirname, '../build/proto');
 
   await fs.remove(proto);
